✨ add upper bar system

+++ src/editor/main.cpp
@@ -1,13 +1,68 @@
 #include <vector>
 
 #include "graph.hpp"
+#include "ui/top_bar.hpp"
 
 int main() {
+    SetConfigFlags(FLAG_WINDOW_RESIZABLE);
     InitWindow(800, 600, "Nebula Test");
-    Vector2 cursor_postion = {0, 0};
+    if (!IsWindowReady()) {
+        return 1;
+    }
+    Vector2 cursor_position = {0, 0};
+    bool should_quit = false;
+    auto request_quit = [&should_quit]() { should_quit = true; };
 
     // Create graph
     core::Graph graph;
+
+    std::vector<editor_ui::Menu> menus;
+
+    menus.push_back(editor_ui::Menu{
+        "File",
+        {
+            editor_ui::MenuItem{"New", [&graph]() { graph = core::Graph{}; }},
+            editor_ui::MenuItem{
+                "Save",
+                []() { TraceLog(LOG_INFO, "Save requested from top bar"); }},
+            editor_ui::MenuItem{
+                "Load",
+                []() { TraceLog(LOG_INFO, "Load requested from top bar"); }},
+            editor_ui::MenuItem{"Quit", request_quit},
+        },
+    });
+
+    menus.push_back(editor_ui::Menu{
+        "Add",
+        {
+            editor_ui::MenuItem{"Variable",
+                                [&graph, &cursor_position]() {
+                                    graph.AddNode(
+                                        core::NodeBase::NodeKind::kVariable,
+                                        {cursor_position.x - 50,
+                                         cursor_position.y - 25});
+                                }},
+            editor_ui::MenuItem{"Literal",
+                                [&graph, &cursor_position]() {
+                                    graph.AddNode(
+                                        core::NodeBase::NodeKind::kLiteral,
+                                        {cursor_position.x - 50,
+                                         cursor_position.y - 25});
+                                }},
+        },
+    });
+
+    menus.push_back(editor_ui::Menu{
+        "Edit",
+        {
+            editor_ui::MenuItem{"Duplicate selection",
+                                [&graph]() { graph.DuplicateSelectedNode(); }},
+            editor_ui::MenuItem{"Delete selection",
+                                [&graph]() { graph.DeleteSelectedNodes(); }},
+        },
+    });
+
+    editor_ui::TopBar top_bar(std::move(menus));
 
     // Create a visual node
     auto node1 = graph.AddNode(core::NodeBase::NodeKind::kLiteral, {500, 100});
@@ -21,7 +76,7 @@ int main() {
 
     SetTargetFPS(60);
 
-    while (!WindowShouldClose()) {
+    while (!WindowShouldClose() && !should_quit) {
         if (IsKeyPressed(KEY_H)) {
             if (IsCursorHidden()) {
                 ShowCursor();
@@ -32,34 +87,45 @@ int main() {
 
         if (IsKeyPressed(KEY_Q)) {
             graph.AddNode(core::NodeBase::NodeKind::kVariable,
-                          {cursor_postion.x - 50, cursor_postion.y - 25});
+                          {cursor_position.x - 50, cursor_position.y - 25});
         }
         if (IsKeyPressed(KEY_W)) {
             graph.AddNode(core::NodeBase::NodeKind::kLiteral,
-                          {cursor_postion.x - 50, cursor_postion.y - 25});
+                          {cursor_position.x - 50, cursor_position.y - 25});
         }
         if (IsKeyDown(KEY_LEFT_CONTROL) && IsKeyPressed(KEY_D)) {
             graph.DuplicateSelectedNode();
         }
 
-        cursor_postion = GetMousePosition();
+        cursor_position = GetMousePosition();
+        top_bar.Update();
 
-        graph.DeleteWithMouse();
-        graph.CheckNodeMovement();
+        if (!top_bar.BlocksGraphInput()) {
+            graph.DeleteWithMouse();
+            graph.CheckNodeMovement();
+        }
 
         BeginDrawing();
         ClearBackground(RAYWHITE);
 
         // Draw visual nodes in the graph
-        graph.SelectWithMouse();
+        if (!top_bar.BlocksGraphInput()) {
+            graph.SelectWithMouse();
+        }
         graph.Draw();
 
-        graph.LinkingWithMouse();
-        graph.HandleContextMenu();
+        if (!top_bar.BlocksGraphInput()) {
+            graph.LinkingWithMouse();
+            graph.HandleContextMenu();
+        }
+
+        top_bar.Draw();
 
         EndDrawing();
     }
 
-    CloseWindow();
+    if (IsWindowReady()) {
+        CloseWindow();
+    }
     return 0;
 }
